updating redundant constants
