+++ faces/client/video/v3io_image.py
@@ -1,10 +1,11 @@
 import cv2
-from config.app_conf import AppConf
 from utils.logger import Logger
 import io
 from datetime import datetime
-
 from video.image_abs import AbsImage
+import base64
+import numpy as np
+import json
 
 
 def get_current_time():
@@ -16,13 +17,13 @@ def get_current_time():
 
 class V3ioImage(AbsImage):
 
-    def __init__(self, logger: Logger, conf: AppConf, frame, camera):
+    def __init__(self, logger: Logger, frame, camera):
         self.logger = logger
-        self.conf = conf
         self.frame = frame
         self.camera = camera
         self.create_time = get_current_time()
-        self.logger.info("Video frame initialize ...")
+        self.image_str = self.convert_frame_to_jpg_str()
+        self.image_json = self.generate_json()
 
     def convert_frame_to_bytes(self):
         is_success, buffer = cv2.imencode('.jpg', self.frame)
@@ -32,7 +33,53 @@ class V3ioImage(AbsImage):
     def get_create_time(self):
         return self.create_time
 
+    def generate_json(self):
+        return json.dumps({"content":self.image_str, "time":self.create_time, "camera":self.camera})
 
+    def convert_frame_to_jpg_str(self):
+        #gray = VideoFrame.convert_frame_to_gray(frame)
+        #resized = cv2.resize(gray,(360,480))
+        jpg = self.convert_to_jpg(self.frame)
+        encoded = self.b64_encode_frame(jpg)
+        utf_decoded = self.decode_as_utf(encoded)
+        return utf_decoded
+
+    @staticmethod
+    def convert_frame_to_gray(frame):
+        return cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
+
+    @staticmethod
+    def b64_encode_frame(frame):
+        return base64.b64encode(frame)
+
+    @staticmethod
+    def convert_gray_frame_to_bytes(frame):
+        return frame.tobytes()
+
+    @staticmethod
+    def decode_frame(frame):
+        return base64.b64decode(frame)
+
+    @staticmethod
+    def decode_as_utf(item):
+        txt_decoded_utf = item.decode('utf-8')
+        return txt_decoded_utf
+
+    @staticmethod
+    def convert_to_jpg(frame):
+        retval, buffer = cv2.imencode('.jpg', frame)
+        if retval:
+            return buffer
+        else:
+            raise
+
+
+    @staticmethod
+    def jpg_str_to_frame(frame):
+        jpg_original = base64.b64decode(frame)
+        jpg_as_np = np.frombuffer(jpg_original, dtype=np.uint8)
+        img = cv2.imdecode(jpg_as_np, flags=1)
+        return img
